fix(tests): Refactor `account_routes` tests into smaller manageable tests
