+++ test/local/routes/account_routes.js
@@ -5,65 +5,17 @@
 var sinon = require('sinon')
 
 var test = require('tap').test
-var mocks = require('../mocks')
-var getRoute = require('../routes_helpers').getRoute
-var proxyquire = require('proxyquire')
+var mocks = require('../../mocks')
+var getRoute = require('../../routes_helpers').getRoute
+var makeRoutes = require('../../routes_helpers').makeAccountRoutes
 
-var P = require('../../lib/promise')
+var P = require('../../../lib/promise')
 var uuid = require('uuid')
 var crypto = require('crypto')
-var isA = require('joi')
-var error = require('../../lib/error')
-var log = require('../../lib/log')
+var error = require('../../../lib/error')
+var log = require('../../../lib/log')
 
 var TEST_EMAIL = 'foo@gmail.com'
-var TEST_EMAIL_INVALID = 'example@dotless-domain'
-
-var makeRoutes = function (options, requireMocks) {
-  options = options || {}
-
-  var config = options.config || {}
-  config.verifierVersion = config.verifierVersion || 0
-  config.smtp = config.smtp ||  {}
-  config.memcached = config.memcached || {
-    address: '127.0.0.1:1121',
-    idle: 500,
-    lifetime: 30
-  }
-  config.i18n = {
-    supportedLanguages: ['en'],
-    defaultLanguage: 'en'
-  }
-
-  var log = options.log || mocks.mockLog()
-  var Password = options.Password || require('../../lib/crypto/password')(log, config)
-  var db = options.db || mocks.mockDB()
-  var isPreVerified = require('../../lib/preverifier')(error, config)
-  var customs = options.customs || {
-    check: function () { return P.resolve(true) }
-  }
-  var checkPassword = options.checkPassword || require('../../lib/routes/utils/password_check')(log, config, Password, customs, db)
-  var push = options.push || require('../../lib/push')(log, db)
-  var metricsContext = options.metricsContext || log.metricsContext || require('../../lib/metrics/context')(log, config)
-  return proxyquire('../../lib/routes/account', requireMocks || {})(
-    log,
-    crypto,
-    P,
-    uuid,
-    isA,
-    error,
-    db,
-    options.mailer || {},
-    Password,
-    config,
-    customs,
-    isPreVerified,
-    checkPassword,
-    push,
-    metricsContext,
-    options.devices || require('../../lib/devices')(log, db, push)
-  )
-}
 
 function runTest (route, request, assertions) {
   return new P(function (resolve) {
@@ -71,155 +23,8 @@ function runTest (route, request, assertions) {
       resolve(response)
     })
   })
-  .then(assertions)
+    .then(assertions)
 }
-
-test('/recovery_email/status', function (t) {
-  t.plan(2)
-  var config = {
-    signinConfirmation: {}
-  }
-  var mockDB = mocks.mockDB()
-  var pushCalled
-  var mockLog = mocks.mockLog({
-    increment: function (name) {
-      if (name === 'recovery_email_reason.push') {
-        pushCalled = true
-      }
-    }
-  })
-  var accountRoutes = makeRoutes({
-    config: config,
-    db: mockDB,
-    log: mockLog
-  })
-  var route = getRoute(accountRoutes, '/recovery_email/status')
-
-  t.test('sign-in confirmation disabled', function (t) {
-    t.plan(2)
-    config.signinConfirmation.enabled = false
-
-    t.test('invalid email', function (t) {
-      t.plan(2)
-      var mockRequest = mocks.mockRequest({
-        credentials: {
-          email: TEST_EMAIL_INVALID
-        }
-      })
-
-      t.test('unverified account', function (t) {
-        mockRequest.auth.credentials.emailVerified = false
-
-        return runTest(route, mockRequest, function (response) {
-          t.equal(mockDB.deleteAccount.callCount, 1)
-          t.equal(mockDB.deleteAccount.firstCall.args[0].email, TEST_EMAIL_INVALID)
-          t.equal(response.errno, error.ERRNO.INVALID_TOKEN)
-        })
-        .then(function () {
-          mockDB.deleteAccount.reset()
-        })
-      })
-
-      t.test('verified account', function (t) {
-        mockRequest.auth.credentials.uid = uuid.v4('binary').toString('hex')
-        mockRequest.auth.credentials.emailVerified = true
-        mockRequest.auth.credentials.tokenVerified = true
-
-        return runTest(route, mockRequest, function (response) {
-          t.equal(mockDB.deleteAccount.callCount, 0)
-          t.deepEqual(response, {
-            email: TEST_EMAIL_INVALID,
-            verified: true,
-            emailVerified: true,
-            sessionVerified: true
-          })
-        })
-      })
-    })
-
-    t.test('valid email, verified account', function (t) {
-      pushCalled = false
-      var mockRequest = mocks.mockRequest({
-        credentials: {
-          uid: uuid.v4('binary').toString('hex'),
-          email: TEST_EMAIL,
-          emailVerified: true,
-          tokenVerified: true
-        },
-        query: {
-          reason: 'push'
-        }
-      })
-
-      return runTest(route, mockRequest, function (response) {
-        t.equal(pushCalled, true)
-
-        t.deepEqual(response, {
-          email: TEST_EMAIL,
-          verified: true,
-          emailVerified: true,
-          sessionVerified: true
-        })
-      })
-    })
-  })
-
-  t.test('sign-in confirmation enabled', function (t) {
-    t.plan(3)
-    config.signinConfirmation.enabled = true
-    config.signinConfirmation.sample_rate = 1
-    var mockRequest = mocks.mockRequest({
-      credentials: {
-        uid: uuid.v4('binary').toString('hex'),
-        email: TEST_EMAIL
-      }
-    })
-
-    t.test('verified account, verified session', function (t) {
-      mockRequest.auth.credentials.emailVerified = true
-      mockRequest.auth.credentials.tokenVerified = true
-
-      return runTest(route, mockRequest, function (response) {
-        t.deepEqual(response, {
-          email: TEST_EMAIL,
-          verified: true,
-          sessionVerified: true,
-          emailVerified: true
-        })
-      })
-    })
-
-    t.test('verified account, unverified session, must verify session', function (t) {
-      mockRequest.auth.credentials.emailVerified = true
-      mockRequest.auth.credentials.tokenVerified = false
-      mockRequest.auth.credentials.mustVerify = true
-
-      return runTest(route, mockRequest, function (response) {
-        t.deepEqual(response, {
-          email: TEST_EMAIL,
-          verified: false,
-          sessionVerified: false,
-          emailVerified: true
-        })
-      })
-    })
-
-    t.test('verified account, unverified session, neednt verify session', function (t) {
-      mockRequest.auth.credentials.emailVerified = true
-      mockRequest.auth.credentials.tokenVerified = false
-      mockRequest.auth.credentials.mustVerify = false
-
-      return runTest(route, mockRequest, function (response) {
-        t.deepEqual(response, {
-          email: TEST_EMAIL,
-          verified: true,
-          sessionVerified: false,
-          emailVerified: true
-        })
-      })
-    })
-  })
-})
 
 test('/account/reset', function (t) {
   var uid = uuid.v4('binary')
@@ -265,293 +70,6 @@ test('/account/reset', function (t) {
     t.equal(args[0], 'account.reset', 'first argument was event name')
     t.equal(args[1], mockRequest, 'second argument was request object')
     t.deepEqual(args[2], { uid: uid.toString('hex') }, 'third argument contained uid')
-  })
-})
-
-test('/account/device', function (t) {
-  t.plan(4)
-  var config = {}
-  var uid = uuid.v4('binary')
-  var deviceId = crypto.randomBytes(16)
-  var mockRequest = mocks.mockRequest({
-    credentials: {
-      deviceCallbackPublicKey: '',
-      deviceCallbackURL: '',
-      deviceId: deviceId,
-      deviceName: 'my awesome device',
-      deviceType: 'desktop',
-      tokenId: crypto.randomBytes(16),
-      uid: uid
-    },
-    payload: {
-      id: deviceId.toString('hex'),
-      name: 'my awesome device'
-    }
-  })
-  var mockDevices = mocks.mockDevices()
-  var mockLog = mocks.spyLog()
-  var accountRoutes = makeRoutes({
-    config: config,
-    devices: mockDevices,
-    log: mockLog
-  })
-  var route = getRoute(accountRoutes, '/account/device')
-
-  t.test('identical data', function (t) {
-    return runTest(route, mockRequest, function (response) {
-      t.equal(mockLog.increment.callCount, 1, 'a counter was incremented')
-      t.equal(mockLog.increment.firstCall.args[0], 'device.update.spurious')
-
-      t.deepEqual(response, mockRequest.payload)
-    })
-    .then(function () {
-      mockLog.increment.reset()
-    })
-  })
-
-  t.test('different data', function (t) {
-    mockRequest.auth.credentials.deviceId = crypto.randomBytes(16)
-    var payload = mockRequest.payload
-    payload.name = 'my even awesomer device'
-    payload.type = 'phone'
-    payload.pushCallback = 'https://push.services.mozilla.com/123456'
-    payload.pushPublicKey = 'SomeEncodedBinaryStuffThatDoesntGetValidedByThisTest'
-
-    return runTest(route, mockRequest, function (response) {
-      t.equal(mockLog.increment.callCount, 5, 'the counters were incremented')
-      t.equal(mockLog.increment.getCall(0).args[0], 'device.update.sessionToken')
-      t.equal(mockLog.increment.getCall(1).args[0], 'device.update.name')
-      t.equal(mockLog.increment.getCall(2).args[0], 'device.update.type')
-      t.equal(mockLog.increment.getCall(3).args[0], 'device.update.pushCallback')
-      t.equal(mockLog.increment.getCall(4).args[0], 'device.update.pushPublicKey')
-
-      t.equal(mockDevices.upsert.callCount, 1, 'devices.upsert was called once')
-      var args = mockDevices.upsert.args[0]
-      t.equal(args.length, 3, 'devices.upsert was passed three arguments')
-      t.equal(args[0], mockRequest, 'first argument was request object')
-      t.deepEqual(args[1].tokenId, mockRequest.auth.credentials.tokenId, 'second argument was session token')
-      t.deepEqual(args[1].uid, uid, 'sessionToken.uid was correct')
-      t.deepEqual(args[2], mockRequest.payload, 'third argument was payload')
-    })
-    .then(function () {
-      mockLog.increment.reset()
-      mockDevices.upsert.reset()
-    })
-  })
-
-  t.test('with no id in payload', function (t) {
-    mockRequest.payload.id = undefined
-
-    return runTest(route, mockRequest, function (response) {
-      t.equal(mockLog.increment.callCount, 0, 'log.increment was not called')
-
-      t.equal(mockDevices.upsert.callCount, 1, 'devices.upsert was called once')
-      var args = mockDevices.upsert.args[0]
-      t.equal(args[2].id, mockRequest.auth.credentials.deviceId.toString('hex'), 'payload.id defaulted to credentials.deviceId')
-    })
-    .then(function () {
-      mockLog.increment.reset()
-      mockDevices.upsert.reset()
-    })
-  }, t)
-
-  t.test('device updates disabled', function (t) {
-    config.deviceUpdatesEnabled = false
-
-    return runTest(route, mockRequest, function () {
-      t.fail('should have thrown')
-    })
-    .catch(function (err) {
-      t.equal(err.output.statusCode, 503, 'correct status code is returned')
-      t.equal(err.errno, error.ERRNO.FEATURE_NOT_ENABLED, 'correct errno is returned')
-    })
-  })
-})
-
-test('/account/devices/notify', function (t) {
-  t.plan(5)
-  var config = {}
-  var uid = uuid.v4('binary')
-  var mockRequest = mocks.mockRequest({
-    credentials: {
-      uid: uid.toString('hex')
-    }
-  })
-  var pushPayload = {
-    isValid: true,
-    version: 1,
-    command: 'sync:collection_changed',
-    data: {
-      collections: ['clients']
-    }
-  }
-  var mockPush = mocks.mockPush()
-  var validate = sinon.spy(function (payload) { return payload.isValid })
-  var mockAjv = function () {
-    return {
-      compile: function () {
-        return validate
-      }
-    }
-  }
-  var sandbox = sinon.sandbox.create()
-  var mockCustoms = {
-    checkAuthenticated: sandbox.spy(function () {
-      return P.resolve()
-    })
-  }
-  var accountRoutes = makeRoutes({
-    config: config,
-    customs: mockCustoms,
-    push: mockPush
-  }, {
-    ajv: mockAjv
-  })
-  var route = getRoute(accountRoutes, '/account/devices/notify')
-
-  t.test('bad payload', function (t) {
-    mockRequest.payload = {
-      to: ['bogusid1'],
-      payload: {
-        isValid: false
-      }
-    }
-    return runTest(route, mockRequest, function () {
-      t.fail('should have thrown')
-    })
-    .catch(function (err) {
-      t.equal(validate.callCount, 1, 'ajv validator function was called')
-      t.equal(mockPush.pushToDevices.callCount, 0, 'mockPush.pushToDevices was not called')
-      t.equal(err.errno, 107, 'Correct errno for invalid push payload')
-    })
-  })
-
-  t.test('all devices', function (t) {
-    mockRequest.payload = {
-      to: 'all',
-      excluded: ['bogusid'],
-      TTL: 60,
-      payload: pushPayload
-    }
-    return runTest(route, mockRequest, function (response) {
-      t.equal(mockCustoms.checkAuthenticated.callCount, 1, 'mockCustoms.checkAuthenticated was called once')
-      t.equal(mockPush.pushToAllDevices.callCount, 1, 'mockPush.pushToAllDevices was called once')
-      var args = mockPush.pushToAllDevices.args[0]
-      t.equal(args.length, 3, 'mockPush.pushToAllDevices was passed three arguments')
-      t.equal(args[0], uid.toString('hex'), 'first argument was the device uid')
-      t.equal(args[1], 'devicesNotify', 'second argument was the devicesNotify reason')
-      t.deepEqual(args[2], {
-        data: new Buffer(JSON.stringify(pushPayload)),
-        excludedDeviceIds: ['bogusid'],
-        TTL: 60
-      }, 'third argument was the push options')
-    })
-  })
-
-  t.test('specific devices', function (t) {
-    mockCustoms.checkAuthenticated.reset()
-    mockRequest.payload = {
-      to: ['bogusid1', 'bogusid2'],
-      TTL: 60,
-      payload: pushPayload
-    }
-    return runTest(route, mockRequest, function (response) {
-      t.equal(mockCustoms.checkAuthenticated.callCount, 1, 'mockCustoms.checkAuthenticated was called once')
-      t.equal(mockPush.pushToDevices.callCount, 1, 'mockPush.pushToDevices was called once')
-      var args = mockPush.pushToDevices.args[0]
-      t.equal(args.length, 4, 'mockPush.pushToDevices was passed four arguments')
-      t.equal(args[0], uid.toString('hex'), 'first argument was the device uid')
-      t.deepEqual(args[1], ['bogusid1', 'bogusid2'], 'second argument was the list of device ids')
-      t.equal(args[2], 'devicesNotify', 'third argument was the devicesNotify reason')
-      t.deepEqual(args[3], {
-        data: new Buffer(JSON.stringify(pushPayload)),
-        TTL: 60
-      }, 'fourth argument was the push options')
-    })
-  })
-
-  t.test('device driven notifications disabled', function (t) {
-    config.deviceNotificationsEnabled = false
-    mockRequest.payload = {
-      to: 'all',
-      excluded: ['bogusid'],
-      TTL: 60,
-      payload: pushPayload
-    }
-
-    return runTest(route, mockRequest, function () {
-      t.fail('should have thrown')
-    })
-    .catch(function (err) {
-      t.equal(err.output.statusCode, 503, 'correct status code is returned')
-      t.equal(err.errno, error.ERRNO.FEATURE_NOT_ENABLED, 'correct errno is returned')
-    })
-  })
-
-  t.test('throws error if customs blocked the request', function (t) {
-    config.deviceNotificationsEnabled = true
-
-    mockCustoms = {
-      checkAuthenticated: sandbox.spy(function () {
-        throw error.tooManyRequests(1)
-      })
-    }
-    route = getRoute(makeRoutes({customs: mockCustoms}), '/account/devices/notify')
-
-    return runTest(route, mockRequest, function (response) {
-      t.fail('should have thrown')
-    })
-    .catch(function (err) {
-      t.equal(mockCustoms.checkAuthenticated.callCount, 1, 'mockCustoms.checkAuthenticated was called once')
-      t.equal(err.message, 'Client has sent too many requests')
-    })
-  })
-})
-
-test('/account/device/destroy', function (t) {
-  var uid = uuid.v4('binary')
-  var deviceId = crypto.randomBytes(16).toString('hex')
-  var mockLog = mocks.spyLog()
-  var mockDB = mocks.mockDB()
-  var mockRequest = mocks.mockRequest({
-    credentials: {
-      uid: uid.toString('hex'),
-    },
-    payload: {
-      id: deviceId
-    }
-  })
-  var mockPush = mocks.mockPush()
-  var accountRoutes = makeRoutes({
-    db: mockDB,
-    log: mockLog,
-    push: mockPush
-  })
-  var route = getRoute(accountRoutes, '/account/device/destroy')
-
-  return runTest(route, mockRequest, function () {
-    t.equal(mockDB.deleteDevice.callCount, 1)
-
-    t.equal(mockPush.notifyDeviceDisconnected.callCount, 1)
-    t.equal(mockPush.notifyDeviceDisconnected.firstCall.args[0], mockRequest.auth.credentials.uid)
-    t.equal(mockPush.notifyDeviceDisconnected.firstCall.args[1], deviceId)
-
-    t.equal(mockLog.activityEvent.callCount, 1, 'log.activityEvent was called once')
-    var args = mockLog.activityEvent.args[0]
-    t.equal(args.length, 3, 'log.activityEvent was passed three arguments')
-    t.equal(args[0], 'device.deleted', 'first argument was event name')
-    t.equal(args[1], mockRequest, 'second argument was request object')
-    t.deepEqual(args[2], { uid: uid.toString('hex'), device_id: deviceId }, 'third argument contained uid and deviceId')
-
-    t.equal(mockLog.notifyAttachedServices.callCount, 1)
-    args = mockLog.notifyAttachedServices.args[0]
-    t.equal(args.length, 3)
-    t.equal(args[0], 'device:delete')
-    t.equal(args[1], mockRequest)
-    var details = args[2]
-    t.equal(details.uid, uid.toString('hex'))
-    t.equal(details.id, deviceId)
-    t.ok(Date.now() - details.timestamp < 100)
   })
 })
 
@@ -1197,166 +715,6 @@ test('/account/login', function (t) {
   })
 })
 
-test('/recovery_email/verify_code', function (t) {
-  t.plan(2)
-  var uid = uuid.v4('binary').toString('hex')
-  var mockRequest = mocks.mockRequest({
-    query: {},
-    payload: {
-      uid: uid,
-      code: 'e3c5b0e3f5391e134596c27519979b93',
-      service: 'sync'
-    }
-  })
-  var dbData = {
-    email: TEST_EMAIL,
-    emailCode: Buffer(mockRequest.payload.code, 'hex'),
-    emailVerified: false,
-    uid: uid
-  }
-  var dbErrors = {
-    verifyTokens: error.invalidVerificationCode({})
-  }
-  var mockDB = mocks.mockDB(dbData, dbErrors)
-  var mockLog = mocks.spyLog()
-  var mockMailer = mocks.mockMailer()
-  var accountRoutes = makeRoutes({
-    checkPassword: function () {
-      return P.resolve(true)
-    },
-    config: {},
-    customs: {
-      check: function () {
-        return P.resolve()
-      }
-    },
-    db: mockDB,
-    log: mockLog,
-    mailer: mockMailer
-  })
-  var route = getRoute(accountRoutes, '/recovery_email/verify_code')
-  t.test('verifyTokens rejects with INVALID_VERIFICATION_CODE', function (t) {
-    t.plan(2)
-
-    t.test('without a reminder payload', function (t) {
-      return runTest(route, mockRequest, function (response) {
-        t.equal(mockDB.verifyTokens.callCount, 1, 'calls verifyTokens')
-        t.equal(mockDB.verifyEmail.callCount, 1, 'calls verifyEmail')
-        t.equal(mockLog.notifyAttachedServices.callCount, 1, 'logs verified')
-
-        t.equal(mockMailer.sendPostVerifyEmail.callCount, 1, 'sendPostVerifyEmail was called once')
-
-        t.equal(mockLog.activityEvent.callCount, 1, 'activityEvent was called once')
-        var args = mockLog.activityEvent.args[0]
-        t.equal(args.length, 3, 'activityEvent was passed three arguments')
-        t.equal(args[0], 'account.verified', 'first argument was event name')
-        t.deepEqual(args[1], {
-          auth: {
-            credentials: {
-              uid: Buffer(uid, 'hex'),
-              id: mockRequest.payload.code,
-            }
-          },
-          headers: mockRequest.headers,
-          payload: mockRequest.payload,
-          query: mockRequest.query
-        }, 'second argument was synthesized request object')
-        t.deepEqual(args[2], {
-          uid: uid.toString('hex')
-        }, 'third argument contained uid')
-
-        t.equal(JSON.stringify(response), '{}')
-      })
-      .then(function () {
-        mockDB.verifyTokens.reset()
-        mockDB.verifyEmail.reset()
-        mockLog.activityEvent.reset()
-        mockLog.notifyAttachedServices.reset()
-        mockMailer.sendPostVerifyEmail.reset()
-      })
-    })
-
-    t.test('with a reminder payload', function (t) {
-      mockRequest.payload.reminder = 'second'
-
-      return runTest(route, mockRequest, function (response) {
-        t.equal(mockLog.activityEvent.callCount, 2, 'activityEvent was called twice')
-        t.equal(mockLog.activityEvent.args[0][0], 'account.verified', 'first call was account.verified')
-        t.equal(mockMailer.sendPostVerifyEmail.callCount, 1, 'sendPostVerifyEmail was called once')
-
-        var args = mockLog.activityEvent.args[1]
-        t.equal(args.length, 3, 'activityEvent was passed three arguments second time')
-        t.equal(args[0], 'account.reminder', 'first argument was event name')
-        t.equal(args[1], mockRequest, 'second argument was request object')
-        t.deepEqual(args[2], {
-          uid: uid.toString('hex')
-        }, 'third argument contained uid')
-
-        t.equal(JSON.stringify(response), '{}')
-      })
-      .then(function () {
-        mockDB.verifyTokens.reset()
-        mockDB.verifyEmail.reset()
-        mockLog.activityEvent.reset()
-        mockLog.notifyAttachedServices.reset()
-        mockMailer.sendPostVerifyEmail.reset()
-      })
-    })
-  })
-
-  t.test('verifyTokens resolves', function (t) {
-    t.plan(2)
-
-    dbData.emailVerified = true
-    dbErrors.verifyTokens = undefined
-
-    t.test('email verification', function (t) {
-      return runTest(route, mockRequest, function (response) {
-        t.equal(mockDB.verifyTokens.callCount, 1, 'call db.verifyTokens')
-        t.equal(mockDB.verifyEmail.callCount, 0, 'does not call db.verifyEmail')
-        t.equal(mockLog.notifyAttachedServices.callCount, 0, 'does not call log.notifyAttachedServices')
-        t.equal(mockLog.activityEvent.callCount, 0, 'log.activityEvent was not called')
-      })
-      .then(function () {
-        mockDB.verifyTokens.reset()
-      })
-    })
-
-    t.test('sign-in confirmation', function (t) {
-      dbData.emailCode = crypto.randomBytes(16)
-
-      return runTest(route, mockRequest, function (response) {
-        t.equal(mockDB.verifyTokens.callCount, 1, 'call db.verifyTokens')
-        t.equal(mockDB.verifyEmail.callCount, 0, 'does not call db.verifyEmail')
-        t.equal(mockLog.notifyAttachedServices.callCount, 0, 'does not call log.notifyAttachedServices')
-
-        t.equal(mockLog.activityEvent.callCount, 1, 'log.activityEvent was called once')
-        var args = mockLog.activityEvent.args[0]
-        t.equal(args.length, 3, 'log.activityEvent was passed three arguments')
-        t.equal(args[0], 'account.confirmed', 'first argument was event name')
-        t.deepEqual(args[1], {
-          auth: {
-            credentials: {
-              uid: Buffer(uid, 'hex'),
-              id: mockRequest.payload.code,
-            }
-          },
-          headers: mockRequest.headers,
-          payload: mockRequest.payload,
-          query: mockRequest.query
-        }, 'second argument was synthesized request object')
-        t.deepEqual(args[2], {
-          uid: uid.toString('hex')
-        }, 'third argument contained uid')
-      })
-      .then(function () {
-        mockDB.verifyTokens.reset()
-        mockLog.activityEvent.reset()
-      })
-    })
-  })
-})
-
 test('/account/keys', function (t) {
   t.plan(2)
   var keyFetchTokenId = crypto.randomBytes(16)
@@ -1469,58 +827,3 @@ test('/account/destroy', function (t) {
     t.equal(args[2].uid, uid.toString('hex'), 'third argument was event data')
   })
 })
-
-test('/account/devices', function (t) {
-  var mockRequest = mocks.mockRequest({
-    credentials: {
-      uid: crypto.randomBytes(16),
-      tokenId: crypto.randomBytes(16)
-    },
-    payload: {}
-  })
-  var unnamedDevice = { sessionToken: crypto.randomBytes(16) }
-  var mockDB = mocks.mockDB({
-    devices: [
-      { name: 'current session', type: 'mobile', sessionToken: mockRequest.auth.credentials.tokenId },
-      { name: 'has no type', sessionToken: crypto.randomBytes(16) },
-      { name: 'has device type', sessionToken: crypto.randomBytes(16), uaDeviceType: 'wibble' },
-      unnamedDevice
-    ]
-  })
-  var mockDevices = mocks.mockDevices()
-  var accountRoutes = makeRoutes({
-    db: mockDB,
-    devices: mockDevices
-  })
-  var route = getRoute(accountRoutes, '/account/devices')
-
-  return runTest(route, mockRequest, function (response) {
-    t.ok(Array.isArray(response), 'response is array')
-    t.equal(response.length, 4, 'response contains 4 items')
-
-    t.equal(response[0].name, 'current session')
-    t.equal(response[0].type, 'mobile')
-    t.equal(response[0].sessionToken, undefined)
-    t.equal(response[0].isCurrentDevice, true)
-
-    t.equal(response[1].name, 'has no type')
-    t.equal(response[1].type, 'desktop')
-    t.equal(response[1].sessionToken, undefined)
-    t.equal(response[1].isCurrentDevice, false)
-
-    t.equal(response[2].name, 'has device type')
-    t.equal(response[2].type, 'wibble')
-    t.equal(response[2].isCurrentDevice, false)
-
-    t.equal(response[3].name, null)
-
-    t.equal(mockDB.devices.callCount, 1, 'db.devices was called once')
-    t.equal(mockDB.devices.args[0].length, 1, 'db.devices was passed one argument')
-    t.deepEqual(mockDB.devices.args[0][0], mockRequest.auth.credentials.uid, 'db.devices was passed uid')
-
-    t.equal(mockDevices.synthesizeName.callCount, 1, 'mockDevices.synthesizeName was called once')
-    t.equal(mockDevices.synthesizeName.args[0].length, 1, 'mockDevices.synthesizeName was passed one argument')
-    t.equal(mockDevices.synthesizeName.args[0][0], unnamedDevice, 'mockDevices.synthesizeName was passed unnamed device')
-  })
-})
-
